Refactor clubs-list.jsx into multiple files
